requirements-dev.txt
-r requirements.txt
django-gulp>=2.3.0
django-livereload>=1.5
django-compressor>=2.0
